Added a mono target build script and corrected issues detected by the mono compiler - using xbuild command

+++ build.ps1
@@ -1,16 +1,19 @@
 # Push current location
 pushd;
 
+# Ensure runtime available
+if (! test-path c:\windows\microsoft.net\framework64\v4.0.30319\msbuild.exe){ write-host 'No compiler ?'; exit 1; }
+
 # Paths
 $rootDirectoryPath = split-path $MyInvocation.MyCommand.Path;
 $outputDirectoryPath = join-path $rootDirectoryPath 'bin';
 
 # Change directories
-pushd;
 cd $rootDirectoryPath;
 
 # Build
-c:\windows\microsoft.net\framework64\v4.0.30319\msbuild.exe src\PMCG.Messaging.Client\PMCG.Messaging.Client.csproj /target:ReBuild /property:Configuration=Release;
+# Could just build the client csproj but want to make sure all projects are good
+c:\windows\microsoft.net\framework64\v4.0.30319\msbuild.exe src\PMCG.Messaging.sln /target:ReBuild /property:Configuration=Release;
 
 # Prepare output directory
 if (test-path $outputDirectoryPath){ rm $outputDirectoryPath\*; } else { mkdir $outputDirectoryPath | out-null; }
